Mise en place des notes, des modifs et suppressions avec confirmation Javascript et modification de la recherche pour l'affichage des résultats plus pertinent

+++ controllers/category.controller.php
@@ -1,111 +1,12 @@
 <?php
 
 class CategoryController
-{
-    function getPageAddImage()
-    {
-        $alert = Security::checkAlert();
-        $title = "Ajout de d'image";
-        $description = "Page permettant l'ajout d'images'";
-
-        $MyBreadcrumb = new MyBreadcrumb();
-        $MyBreadcrumb->add('Ajout d\'images', '#');
-        $breadcrumb = $MyBreadcrumb->breadcrumb();
-
-        if(Security::checkAccess())
-        {
-            if(!empty($_FILES))
-            {
-                try
-                {
-                    $nb_files = count($_FILES['img_file']['name']);
-
-                    for($i=0; $i<$nb_files; $i++)
-                    {                        
-                        $filename =  explode('.', $_FILES['img_file']['name'][$i]); 
-                        // 0-name 1-extension
-                        $name = cleanString($filename[0]);
-                        $tempName = $name;
-                        $j = 1;
-                        // if image name exists, change his name as long as needed
-                        while(getIfImageExist($tempName,$_SESSION['user']['id']) > 0)
-                        {
-                            $tempName = $name.$j;
-                            $j++;
-                        }
-                        $fileImage['name'] = $_FILES['img_file']['name'][$i];
-                        $fileImage['type'] = $_FILES['img_file']['type'][$i];
-                        $fileImage['tmp_name'] = $_FILES['img_file']['tmp_name'][$i];
-                        $fileImage['error'] = $_FILES['img_file']['error'][$i];
-                        $fileImage['size'] = $_FILES['img_file']['size'][$i];
-
-                        // create folder if not exist
-                        $dir = USER_DIRECTORY."/images";
-                        if(!file_exists($dir)) mkdir($dir,0777);
-                        $directory = $dir."/user".$_SESSION['user']['id']."/";
-                        
-                        $tmp_name_img = cleanString($tempName);
-                        $imgName = addImg($fileImage, $directory,$tmp_name_img);
-                        
-                        $description = "Image representant ".$name;
-                        $id_image =  insertImageIntoBD($tempName, $imgName, $description, $_SESSION['user']['id']);
-                        }
-                }
-                catch(Exception $e)
-                {
-                    throw new Exception("Erreur lors de l'insertion de l'image");
-                }
-
-                if($nb_files > 1)
-                    $alert['msg'] = "Les images ont été ajoutées";
-                else
-                    $alert['msg']  = "L'image \"".$filename[0]."\" a été ajoutée";
-
-                $alert['type'] = ALERT_SUCCESS;
-                Alert::setAlert($alert);
-                header ('Location: library');
-                return;
-            }
-            require_once "views/back/images/addImage.view.php";
-        }
-        else 
-        {
-            throw new Exception("Acces interdit si vous n'êtes pas authentifié");
-        }
-    }
-
-
-    function getPageDeleteImage()
-    {
-        $alert = Security::checkAlert();
-        $title = "Supprimer image";
-        $description = "Page de suppression d'image";
-
-        if(isset($_GET['del']))
-        {
-            $id_image = Security::secureHTML($_GET['del']);
-            
-            try
-            {
-                $image = getImageFromID($id_image, $_SESSION['user']['id']);
-                if(deleteImage($id_image,$_SESSION['user']['id'])<1)
-                {
-                    throw new Exception ("Aucune image supprimée");
-                }
-                $url = USER_DIRECTORY."images/user".$_SESSION['user']['id']."/".$image['url'];
-                deleteFile($url);
-                $alert['msg'] = "La suppression de l'image est effective";
-                $alert['type'] = ALERT_WARNING;
-            } catch(Exception $e){
-                $alert['msg'] = "La suppression de l'image n'a pas fonctionné";
-                $alert['type'] = ALERT_DANGER;
-            }
-            Alert::setAlert($alert);
-        }
-        header ('Location: library');
-    }
-
-
+{    
+    /**
+     * getPageCategory
+     *
+     * @return void
+     */
     function getPageCategory()
     {
         $alert = Security::checkAlert();
@@ -128,11 +29,10 @@ class CategoryController
             {
                 $id_user = $_SESSION['user']['id'];
                 $pageNum = (!empty($_GET['pageNum']) ? Security::secureHTML($_GET['pageNum']) : 1);
-                $nb_max_pages = getNbPagesFromCategory($pageNum, $id_category, $id_user);
+                $ResultNumber = getResultNumberFromCategory($pageNum, $id_category, $id_user);
                 $notes = getNotesFromCategory($pageNum, $id_category, $id_user);
                 $category = getCategoryFromID($id_category, $id_user);     
                 
-                // Replace [Library] by <img> in BDD
                 $i = 1;
                 foreach( $notes as $key => $note )
                 {
@@ -170,11 +70,16 @@ class CategoryController
         } 
         else 
         {
-            throw new Exception("Acces interdit si vous n'êtes pas authentifié");
+            throw new Exception("Accès interdit si vous n'êtes pas authentifié");
         }
     }
 
-
+    
+    /**
+     * getPageCategories
+     *
+     * @return void
+     */
     function getPageCategories()
     {
         $alert = Security::checkAlert();
@@ -200,7 +105,12 @@ class CategoryController
         }
     }
 
-
+    
+    /**
+     * getPageAddCategory
+     *
+     * @return void
+     */
     function getPageAddCategory()
     {
         $alert = Security::checkAlert();
@@ -274,10 +184,15 @@ class CategoryController
         }
         else 
         {
-            throw new Exception("Acces interdit si vous n'êtes pas authentifié");
+            throw new Exception("Accès interdit si vous n'êtes pas authentifié");
         }
     }
-
+    
+    /**
+     * getPageEditCategory
+     *
+     * @return void
+     */
     function getPageEditCategory()
     {
         $alert = Security::checkAlert();
@@ -366,16 +281,13 @@ class CategoryController
                 $alert['msg'] = "Catégorie modifiée avec succès";
                 $alert['type'] = ALERT_SUCCESS;
                 Alert::setAlert($alert);
-                // var_dump($alert);
-                // return;
                 header ('Location: category&id='.$id_category);
                 return;
             }
             else
             {
                 if( isset($_POST['category_name']) && empty($_POST['category_name']))
-                {           
-                    // TODO : erreur pour chaque champs
+                {
                     $alert['msg'] = "La categorie ne peut rester vide";
                     $alert['type'] = ALERT_DANGER;
                     Alert::setAlert($alert);
@@ -388,39 +300,48 @@ class CategoryController
             throw new Exception("Accès refusé");
         }
     }
-
+    
+    /**
+     * getPageDeleteCategory
+     *
+     * @return void
+     */
     function getPageDeleteCategory()
     {
-        $alert = Security::checkAlert();
         $title = "Supprimer notes";
         $description = "Page de suppression de notes";
 
-        if(isset($_GET['del']))
+        if(Security::checkAccess())
         {
-            $id_category = Security::secureHTML($_GET['del']);
-            
-            try
+            if(isset($_GET['del']))
             {
-                $image = getImageFromCategory($id_category, $_SESSION['user']['id']);
-                var_dump($image);
-                if(deleteCategory($id_category,$_SESSION['user']['id'])<1){
-                    throw new Exception ("la suppression n'a pas fonctionné en BD");
+                $id_category = Security::secureHTML($_GET['del']);
+                
+                try
+                {
+                    $image = getImageFromCategory($id_category, $_SESSION['user']['id']);
+                    if(deleteCategory($id_category,$_SESSION['user']['id'])<1){
+                        throw new Exception ("la suppression n'a pas fonctionné en BD");
+                    }
+                    // return;
+                    $url = USER_DIRECTORY."icons/user".$_SESSION['user']['id']."/".$image['url'];
+                    deleteFile($url);
+                    $alert['msg'] = "La suppression de la catégorie est effective";
+                    $alert['type'] = ALERT_WARNING;
+                } 
+                catch(Exception $e)
+                {
+                    $alert['msg'] = "La suppression de la catégorie n'a pas fonctionné";
+                    $alert['type'] = ALERT_DANGER;
                 }
-                // return;
-                $url = USER_DIRECTORY."icons/user".$_SESSION['user']['id']."/".$image['url'];
-                deleteFile($url);
-                $alert['msg'] = "La suppression de la catégorie est effective";
-                $alert['type'] = ALERT_WARNING;
-            } 
-            catch(Exception $e)
-            {
-                $alert['msg'] = "La suppression de la catégorie n'a pas fonctionnée";
-                $alert['type'] = ALERT_DANGER;
+                Alert::setAlert($alert);
             }
-            Alert::setAlert($alert);
+            header ('Location: categories');
         }
-        header ('Location: categories');       
-
+        else
+        {
+            throw new Exception("Accès refusé");
+        }
     }
 
 }
